pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "archimedes"
version = "0.1.0"
description = "Archimedes Phase 1 bootstrap"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "alembic>=1.13,<2.0",
  "httpx>=0.27,<1.0",
  "pydantic>=2.7,<3.0",
  "temporalio>=1.8,<2.0",
  "pytest>=8.0,<9.0",
  "python-dotenv>=1.0,<2.0",
  "psycopg[binary]>=3.1,<4.0",
]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["backend*"]
